feat(resource): add edit page for resources
