+++ features/resources/containers/index.ts
@@ -1,2 +1,1 @@
 export { default as ResourceReviewForm } from "./ResourceReviewForm";
-export { default as UpdateResourceForm } from "./UpdateResourceForm";
